refactor : 코드리뷰 기반 리팩터링 - 출력과 입력 클래스 분리
